Show new tool links in the CIDR calculator and use classes instead of IDs
Add a new toollink that provides links to useful tools for ranges
generated by the CIDR calculator. This is shown just underneath
and by default links to a WHOIS tool, a link to CheckUser to check
the range and a link to the CheckUser log for the range. Also convert
the IDs used in the CIDR calculator to classes as .less rules mean
classes can only be used for CSS styling. This also means that,
with some other changes, multiple CIDR calculators can be shown
and work on the same page.

Bug: T311020
Change-Id: Id903efad721c7ed8dc0892216d853e93ade0f33e