Revert some triple equal checks due to incompatabilities on sqlite
Due to unforseen issues, sqlite seems to have cuc_type be in the
$row passed to formatRow as a string value. This means that when
comparing cul_type with a triple equals comparison it will always
fail when using sqlite as PHP won't attempt to convert the type.
Revert back to using double equals so that for sqlite the
comparison will still work. Still a little unsure why this is the
case but I suspect it's because the default for cuc_type in sqlite
DBs created before the abstract schema was created is the string '0'.

Bug: T315224
Change-Id: I40bceb4336b1cadbe8f6a179d2d51e333d7e4d49