Move LogPager.php to the CheckUser pagers as it's only used by CheckUserLog
Move LogPager.php to the CheckUser/Pagers folder as it's a pager
that is only used by the SpecialCheckUserLog.php file which is
located in the CheckUser folder. Also rename it to CheckUserLogPager.php
as this name fits better with the other pagers in the same folder and
makes the distinction that this is only used by the CheckUserLog.

Change-Id: Iee1c88a32b8ab753787bb86169d657349273ccbb