Import CheckUserHelper.js into the extension with major modifications
Import Ladsgroup's CheckUserHelper.js into the extension. This has
been modified in a major way to fit with OOUI where possible and
also to show the XFF information that may be associated with an IP.
The file is not a re-write and permission has been gained from
Ladsgroup and Huji (the two contributors to this tool) to publish
under GPLv2 to allow license compatibility with wikimedia.

Bug: T311378
Change-Id: Id0dd0866bafb79f5ae53fa53d76edb9369f1b5b6