Batch process and format revision comments using preprocessResults
Batch process and format revision comments in the 'get edits'
check type by placing the code in the preprocessResults function
and using CommentFormatter's createRevisionBatch(). Also replace
the throwing of the Exception with a logger warning and in the
case where the revision ID is not found just display the empty
string as the comment.

This also has the nice side effect that cu_changes rows with
a oldid will show the grey and strike when the comment is hidden
even if the comment is still shown to the CheckUser when they have
the needed permissions.

Bug: T234239
Change-Id: I5b5b4a891f304cfbab1549592a8677c061366cb5