Use SelectQueryBuilder in CompareService
Also remove the public method getQueryInfoForSingleTarget, which
is mostly no longer useful for SelectQueryBuilder, except for
building the WHERE conditions. Replace it with a private method,
buildCondsForSingleTarget, which does this.

Remove testGetQueryInfoForSingleTarget, which was testing the
removed functionality. Do not add a test for
buildCondsForSingleTarget, since this is tested via
testGetTotalEditsFromIp and testGetQueryInfo.

Bug: T316433
Change-Id: Id5aa26b3ce112d3fb01140e871bc70e1dab563c3