Show users already tagged or blocked with same params in CU block form
Show users as having been blocked who are already blocked when
reblocking is enabled and the block that was going to be applied was
the same (as determined by DatabaseBlock::equals). Similarly, when a
user is tagged and the user / user talk page already has the same
content, then mark the tag as a success.

To assist in fixing errors after the use of the form, the users
who were blocked and the users who were tagged are now shown
separately after the form is used. This is done as a user could
have been blocked but not tagged (for example the CU can block
users but does not have the ability to edit the user page).

Tests for the code modified in this patch are introduced in a
later patch to keep the size delta per patch manageable.

Bug: T313196
Change-Id: Ie2b7d0bacbe769c0dccb0c8b220b9c7ad920af99