Remove is blocked check as this is done when calling placeBlock()
Remove the check in the doMassUserBlockInternal() function in
SpecialCheckUser.php for whether the account is already blocked.
This is because BlockUser::placeBlock already checks for whether
the user is blocked. This fixes T314510 where the CheckUser block
form did not block IPs if they were already autoblocked because
the autoblock was considered a block by the SpecialCheckUser code
but is ignored by the BlockUser code (and thus Special:Block) for
determining whether a block already exists.

Bug: T314510
Change-Id: I929f62b5b560182444d9127cc4aa39e968861541