Add check links to user tool links on Get edits
Add the check link present on 'Get users' to the 'Get edits' results
by adding it inside the user tool links in the results. This is done
via hooking into a hook that allows CheckUser to add the check link
on only Special:CheckUser. All uses of the user links in CheckUser
would benefit from having the check link included.

As the link is added to the tool links, remove the existing link
from the 'get users' page as this would then be a duplicate.

Bug: T311377
Change-Id: I53d519ba6bbb647d5fedd9b11528e2b23e094aa5