Start on selenium tests for browser side testing of CheckUser
Start on some basic selenium tests for CheckUser to get the extra
config files needed set-up and some basic tests in to test that
it's setup correctly. These tests, unlike phpunit tests, are much
better at testing the browser output. This makes them useful for
testing code and how the code interacts with MediaWiki in general.

Most of the files are from https://www.mediawiki.org/wiki/Selenium/
How-to/Create_the_first_test_in_a_repository but modified to suit
CheckUser.

A method to login as a user with the checkuser / checkuserlog rights
will needed to fully test using selenium.

Bug: T318448
Change-Id: I4c72cc2dec7170e53715375b815e229bae7989a0