Use PagerNavigationBuilder over overriding IndexPager::makeLink
Use the newly introduced PagerNavigationBuilder in 1.39 in CheckUser
as overriding IndexPager::makeLink is being deprecated in favour
of this builder. To do this a new class that extends the builder
then overrides the ::makeLink method to continue to allow the paging
links added in 1.39 to be based on POST requests with a token.

Because the builder was added in 1.39, this change could only be used
for versions of CheckUser that require MediaWiki 1.39 or higher. As
such the extension.json file is updated to note this need.

No functionality of the paging links in CheckUser should change in this
commit, and from my testing this is the case.

Bug: T317914
Change-Id: I22683b39ea390e8eae39e58b41c7fc8ca9f24714