Add block reason dropdown to Special:CheckUser's checkuserblock form
Add block reason dropdown in a similar manner to Special:Block's
dropdown, but instead using a different MediaWiki interface page
to define the dropdown items as the items wanted are likely to be
different.

At the moment the selectandother OOUI field when marked as required
leads the input box to need to have content. As the dropdowns could
be used without any other reason defined, such as when using a
generic summary like "{{CheckUserBlock}}", having the input field
being empty is a allowed use case. This means that until the required
OOUI param can be customised to check that either content is in the
field or a dropdown item is selected, the client side visual
representation of the field being required cannot be used. The server
side check for an empty block reason still is in place, so no blocks
will be made unless a dropdown is chosen or text is in the input field.

This can also be implemented for Special:Investigate, but as there
exists no server-side separate check for an empty reason the block
reason field therefore has to stay required through OOUI. This means
this dropdown can't be implemented yet for Special:Investigate.

Bug: T24120
Change-Id: I9610ceb5435ec2de777ea47b594c95e1ec8b0290