Ensure the limit check is run properly on Special:CheckUser
Due to a oversight in a previous patch, the check for a invalid
range was not performed until after the check for whether the
query should continue. This meant any over limit query still had
a check run but also did not have any IP conds as they were invalid.
This will need to be backported into REL1.39.

For this fix the "skipQuery" boolean is removed and instead the check
is run before the pager object is created. This also adds a new
message key that is shown when the target is outside the allowed limits.

Bug: T318487
Change-Id: I522ff87aae1d494e8412460f21190f042080cf3d
(cherry picked from commit b3e59d153aea6fac28326a27ae610a2454dc44de)