Add multiple integration tests for Hooks.php
Re-adds tests added in I3cde4481cdec1f7bed1358d692258647e0545731
which were later reverted due to the 'cuc_private' field not
existing on all wikis. This patch does not include the change
that lead to the DB errors in T321041 (and does not modify any
files but test files).

Bug: T315648
Change-Id: Ib253381d77656a9e1db5103320936534a1631a72