Reduce code duplication, test, and assign class to each results type
Reduce code duplication by moving code where possible to the
abstract pager. While doing this also add a class to the div
element that wraps around the results in the format "mw-checkuser-
checktype-results" with checktype replaced with "get-ips" et. al.
Also don't show the list toggle on Get users when there are no
results. As part of the deduplication also rename noMatchesMessage
to getEmptyBody to remove the need to call it from the non-abstract
pagers. Also included is the removal of the abstract pager's
reallyDoQuery implementation as this is a direct copy of the parent
class' implementation in IndexPager after the removal of the
skipQuery check.

Apart from the new classes, removing the list toggle when no results
are shown and fixing un-matched trailing </ul> tag there should
be no changes to the output.

These changes also have the benefit of increasing test coverage.

Also add some tests so that AbstractCheckUserPager::__construct
is fully tested, specifically about testing the limit chosen by
the pager given the limit submitted in the form and the maximum
limit defined in the config var.

Bug: T201154
Bug: T311354
Change-Id: I6e88ccb565da3d8103e7fbcfac965d401ec5a0ec