Test HookRunner and provide the UserIdentity for the insert hook
Test HookRunner.php using a very similar method to AbuseFilter's
test for their hook runner code. Also pass the UserIdentity to code
that hooks onto CheckUserInsertChangesRow so that they can view the
user ID and username associated with the row to be inserted as this
may not always be set as cuc_user and cuc_user_text.

Change-Id: I28acaaebd2d0067b700da0930e7b7ba924fa5c1c