Fix the wording used by the checkuser-reset-action message
Fix the wording used by the checkuser-reset-action message as this
is used to indicate when a user successfully causes a password
to be sent by email to the user. This is either triggered by using
Special:PasswordReset or when creating an account which has the
password emailed to the user.

Currently the message incorrectly implies that the password was
reset and changed. To do this the user has to log into using the
password emailed to them. All a user has to do to trigger this
message for a registered account is have the username of the
account (and possibly also the email address depending on settings).

Bug: T322703
Change-Id: Ic5260d62af756e496c852e1c2ab5f66d0ecfa7dc