Enforce the minimum length for the tags as 3 on the client-side
Enforce the minimum length for the talk page tag and the user page
tag as 3 characters in the checkuser 'get users' mass block form
using the minlength attribute. If supported by the browser, this
attribute will cause the browser to display a warning that the
number of characters must be 3 or more.

Currently the code will not add the tags if the length of the wikitext
is less than 3 characters. Adding this check on the client-side
reduces confusion as the current code would lead to no indication
that the tags were not added if the form was otherwise successfully
submitted.

The minlength attribute will not warn the user if there is nothing
inside of input for the tags. It will warn if the checkbox is not
ticked but some content less than 3 characters is present inside
the input box. Avoiding this would require javascript.

Bug: T316441
Depends-On: I30d54f44d2a164615bfd248322eaa4f12270c9b1
Change-Id: I650727a25339aa39bb95a0e545083a1ff2456786