Add tables cu_log_event and cu_private_event for storing log events
Add tables cu_log_event and cu_private_event to the database for use
by CheckUser. These will be used to store entries that originally
went into cu_changes that were for log events. This allows CheckUser
to reference a log ID, store the action text as log parameters and
other changes that come with being able to start with a new table.

The first new table cu_log_event will store log events that have
an associated log ID in the logging table. This table will store
fields required for indexing, checkuser specific fields, the log ID,
and a primary key ID. All other attributes will be got by a JOIN
with the logging table to reduce redundancy.

The second new table, cu_private_event, will store log events that
have no associated log ID in the logging table. This table is created
so that other attributes (such as log parameters) which are not
relevant to edits can be stored while keeping cu_changes from having
columns that are unused when the action is not a log type.

Further tables are likely to be introduced in the future, but these are
all that is needed to address the (many) parent tasks which make this
a high priority.

Bug: T324907
Change-Id: Ied5b0ccbfd73751f9a47674f2e7e98efa469d568