Convert last few CheckUser provided hooks to the HookRunner system
Convert the hooks CheckUserInsertForRecentChange and
SpecialCheckUserGetLinksFromRow to use the HookRunner system. This
includes creating interface files, adding them to the HookRunner
and updating calls to run the Hooks to use the HookRunner file.

This also removes a deprecated call to the static method Hooks::run.

Also fix the documentation for CheckUserInsertChangesRow which
incorrectly specified a since 1.41.

Bug: T312668
Change-Id: Iccfdcaf7c3ddce1da7dc68041955eab867bee090