Split maybePruneIPData and then test newly created pruneIPData
Split maybePruneIPData such that all but the call to mt_rand is
split into a new method named pruneIPData. This is because it is
not possible to easily mock mt_rand, so testing the rest of the
code in that method needs to be instead done in a method that will
always perform the delete operations.

Then create tests to test the pruneIPData method that test for both
deletion of the expired records and keeping the unexpired records.

Bug: T315648
Change-Id: I87fa145cb7718cf1961245da0183e5916a81b24a