Change cu_changes comment migration to write both
259 is the value of SCHEMA_COMPAT_READ_OLD | SCHEMA_COMPAT_WRITE_BOTH

I would like to add a script in a next patch that will populate the
values in the cuc_comment_id column. This script will be added to the
onLoadExtensionSchemaUpdates hook. Since it is a LoggedUpdateMaintenance
script, it will only run once as long as update.php --force is not run
manually. Therefore we should set the migration setting to _write_both
before adding the script, to prevent that there are rows in the db in
which cuc_comment_id is 0.

Bug: T233004
Depends-On: I427f5b800c3cb21a2b221f43bd099fc623bcd9d2
Change-Id: Id016cc68d8b48edf1a5173beb1cbb015683d1f54