Add methods to insert into cu_private_event and cu_log_event
Add methods to Hooks.php to insert rows into cu_private_event and
cu_log_event. To do this new hooks need to be defined by CheckUser
that compliment CheckUserInsertChangesRow, but are specific to each
of the new tables. This means that users of CheckUserInsertChangesRow
will need to also hook into these two new hooks if the actions are to
be applied to log entries.

The new hooks, and the existing hook, will also provide the
associated RecentChange object if a RecentChange triggered the
row insertion. If the row has no RecentChange association, this
is set to null.

The new methods are unused, but are tested. The usage of these
methods will come in future patches so that the change can be kept
smaller and easier to review.

Bug: T324907
Depends-On: I44e54a3fca5558a1cb8d8f06a3990ded863454bc
Change-Id: I7c7754323ade9a8d96273c1742f30b1b5fbe5828