Pager: clarify which table the 'use index' option is for
Selenium tests are currently unhappy when we set the migration config
var to read new, see [1]. No idea why this did not cause issues while
manually testing.

The problem is that if we read from cuc_actor, the actor table joins,
so rdbms no longer knows which table the 'use index' option is for.

[1] - https://phabricator.wikimedia.org/P42933

Bug: T233004
Change-Id: I96da89e19ce1057fd6df31a83f98a8f4b1ddde85