Add write new support for cu_private_event and cu_log_event in Hooks
Add code to Hooks.php that adds WRITE_NEW support for the
CheckUserEventTablesMigrationStage config (writing to cu_log_event
and cu_private_event).

WRITE_NEW and WRITE_OLD must never be set
together as the entry either must be written to one of the new
tables or to the cu_changes table. If they are both set then for
one event two rows would be generated, and there is no way yet to
determine if an entry is duplicated. Maintenance scripts will be
made to allow moving entries between the tables incase write new
needs to be reverted.

The CheckUserInsertForRecentChange hook is also marked as
deprecated with the alternative being the other hooks starting
with "CheckUserInsert". This is because this now deprecated hook
allows the caller to update the row with column names, which means
that it can only be run for inserts to cu_changes. Having new hooks
for inserts using a RecentChange seems unecessary over just using
the hooks called before inserting into the three tables respectively.

The added code is integration tested, and the tests for WRITE_OLD
can be removed once the migration has been fully completed.

Bug: T324907
Change-Id: Ibe1f0a165c9cc46f2f7c45090b77a682f56dc4d5