Add buttons for showing IP addresses next to temporary account user links
* Add links on all special pages except the CheckUser ones (where the IP
  addresses are already visible alongside temporary account names)
* Add links on history and diff pages
* Only add links if the user has the right permissions

Buttons are added via a JS module that looks for the `mw-tempuserlink`
class and passes through data stored in the link HTML, e.g. mw-revid.

When the button is clicked, an API call is made to the TemporaryAccount
endpoints, and the button is replaced with the relevant IP address if
found, or a message indicating that it was not found.

Bug: T326415
Change-Id: I6b98a2f9a59922eb923bad8c39cad05514b5fc66