Add read new support for cu_log comment ID columns
Add read new support for cul_reason_id in Special:CheckUserLog.
As part of this the query has been changed to make the actor table
be directly JOIN'ed as without this the SQL query fails when
trying to join the comment table.

This method is based on how recentchanges interacts with actor and
comment.

Bug: T233004
Change-Id: I0aaea08c1b424511199846535d137f947ca43c7e