Add reason search in CU log and move plaintext gen. code to a service
Add the ability to search for log entries by their reason.
When the reason comment table migration config is set to read new:
 Convert the reason search term to plaintext and match this against
 the stored plaintext reason (using the comment table).
When the reason comment table migration config is set to read old:
 Search for matches to cul_reason using both the plaintext
 version and non-modified version of the reason search term.

This commit does not add wildcard searching to the log as this
made the query too expensive in testing, especially for reasons
that are not used by any log entry. This may be done later when
an alternative method is found that does not make the query too
expensive.

Also as part of this change the code that generates the plaintext
reason is moved to the CheckUserLogService to de-duplicate the code
used to generate it and also prevent changes being not synced over
all definitions.

Bug: T16699
Change-Id: Id08ad02c89a370d3cda659e8b2713ff14f4d78f1