Add read new support for cu_log comment ID columns
Add read new support for cul_reason_id in Special:CheckUserLog.
As part of this the query has been changed to make the actor table
be directly JOIN'ed as without this the SQL query fails when
trying to join the comment table.

This patch uses a different method to the now reverted patch that
added read_new support which is based on the method used to migrate
cu_changes to use the comment table. I have manually tested creating
a log entry with comment IDs of 0 and verified it shows up with
WRITE_BOTH and READ_OLD.

Bug: T233004
Change-Id: I3114f9d26f553af6ff56cbd5b802adfd0a8fd20a