Add CheckUserUnionSelectQueryBuilder and associated factory
Add a CheckUser specific query builder that extends the standard
SelectQueryBuilder to provide a way to generate SQL statements
that select data from cu_changes, cu_log_event and cu_private_event.

This comes with an associated factory that is used to make a new
instance. This factory is registered as a service provided by
CheckUser. It also comes with a helper class that allows the query
builder to modify the lastAlias property of the SelectQueryBuilder's
used to create the sub queries.

This query builder will be used by all parts of CheckUser that
intend to read from the three tables specified above. If reading
from one of the three tables it is better to use SelectQueryBuilder
or the currently used method as this query builder makes
modifications to allow reading from three tables that are not needed
when reading from one.

This query builder is currently unused but is tested, and will be
introduced into use in future commits that will use this query builder
if read new for the event table migration is set.

Bug: T328988
Change-Id: I1d148e18d81d1122838b9dc7e994190f4e796ddf