Remove the database update from 1.39 to remove the cul_reason default
Remove the database update from 1.39 that removes the default value
for cul_reason as the removal and then re-addition of a default value
does not work and causes database errors in postgres installations.

This was found through testing the removal of this update, but I am
not yet sure of a cause. This likely has something to do with batching
the database updates and specifically only issues with postgres. This
patch fixes the immediate problem but likely not the root cause.
However, the root cause doesn't need to be addressed for the purposes
of CheckUser.

Bug: T328483
Change-Id: I4e50f534c2adf7f3c58f9c802b842a85962c80f9