Selenium: Make ::open() async to wait until page fully loaded
Make the open methods for the CheckUserPage and CheckUserLogPage
selenium page classes be async so that tests can wait until the
page has actually loaded. The screenshots from the failed test
indicated that the state of the DOM was not ready or that the test
ran when still on Special:UserRights (to self-grant CheckUser).

This should address the problem, but as this is a rare failure
(presumably caused by slower page loading) it will not be clear if
this has fixed the root cause of the problem.

Bug: T328562
Change-Id: I5728e7a11e8c5cf67a51acad923fa13faf13cb7e