Remove cul_reason column from cu_log
Remove cul_reason column from cu_log as the comment table migration
is now complete.

To achieve this the populateCulComment.php maintenance script needed
some extra code to handle the CheckUser services not being defined.
Because of this tests have been created to ensure that the
maintenance script will work as this change also removes the column
that the data is migrated from.

Bug: T233004
Change-Id: Ib4ac38477b612259e26443d5c70ac005e4544aaf