Add config to control default collapse state of CheckUserHelper
Add a config that allows control of the default collapse state of
the CheckUserHelper fieldset that is added to Special:CheckUser.
Added to make it possible to hide by default the tool on loginwiki
as it is deemed not useful enough to show by default.

Also while I'm modifying that code, add tests to the code that
generates the fieldset that is displayed in Special:CheckUser that
is used by the CheckUserHelper.js script.

Bug: T328726
Change-Id: Ie019fd6ad95d47d51d2ba738f82f69b4ca679f42