Handle rare case that cul_reason_id is NULL
Handle the rare case that cul_reason_id is NULL. This is caused by
pulling the latest master of the CheckUser extension and then
running update.php between 4th June 2022 and 6th June 2022
(which is 17e84c0 to it's reversion in 055744b). Without
this fix the code would fail without migrating the cul_reason field
and then would drop cul_reason loosing the data.

The scope of this issue is limited due to the timescales needed to
cause this. However, it is still important for this to be fixed.

Bug: T328849
Change-Id: Ief2d119f6d0c8c663878d55fff7669acf147d36b