Don't drop cul_actor default if cul_reason is missing for SQLite
Don't run the SQL file to drop the cul_actor default if cul_reason
does not exist for SQLite DBs. This is because the SQLite SQL file
uses a temporary table method that requires cul_reason to exist.
Because modifyExtensionField does not check for this problem, the
SQL file can be run when the state of the DB has cul_reason removed
but nothing preventing this update in the updatelog. This can happen
when running update.php with CheckUser installed for the first time.

Bug: T328851
Change-Id: I2c5e871d66140c6c58c0fae239052f02e5bc3c13