Don't run postgres updates on specific columns unless the field exists
Don't run the postgres updates from 1.39 that are specified using
DatabaseUpdater::addExtensionUpdate for cuc_user, cuc_user_text,
and cul_user if those fields do not exist. Running update.php when
any of these columns are removed results in an error as the method
used requires that the fields exist.

Also fix PopulateCulCommentTest::addDBDataOnce to respect the database
type when creating cul_reason column in cu_log for the test. This test
does not work for postgres, so it is skipped for postgres.

Bug: T328855
Change-Id: Ifbae4e58891b25e5bc10d81afeed95865a9d62b0