Add cuc_only_for_read_old to cu_changes to assist in migration
Add cuc_only_for_read_old to the cu_changes table to assist the
migration of log events out of cu_changes and into the cu_log_event
and cu_private_event tables. This will be used to mark a row as
also being defined in cu_log_event and cu_private_event, and as
such should only be displayed when showing results when reading old.

Without this column there would be no way for the CheckUser
extension to know if a row for a log event also exists in either the
cu_log_event or cu_private_event tables.

Bug: T328874
Change-Id: Id90ae981eacb7695171497089c3b182e1fb9fdcb