Fix action name for password reset email event
Rename the subtype of the cu_private_event log created when a
password reset email is triggered for a user. The name of the action
incorrectly implies that the a reset action has been performed
(i.e. a password was reset) whereas only a password reset email
has been sent which usually only requires a third-party to have a
username or email address associated with the account.

Bug: T324907
Change-Id: I687c0f942464ac5e9bb507534b30717bf15f0c71