Perform insertion to cu_log post send
Perform the insertion to cu_log in CheckUserLogService::addLogEntry
post send so that the TransactionProfiler doesn't raise warnings
about inserts being performed pre-send on a GET request.

Because the logging to cu_log must work (otherwise checks are not
auditable), a critical error is logged if there is a database
error that prevents the insertion (as the step below of using
'error' states that this is intended for monitoring errors).

Bug: T328819
Change-Id: I568392dd62573b8bc45e2130ea5e9026388f3e23