Use translated version of the XFF prefix used by checkUserHelper.js
Create two new messages to replace the untranslated text used by
the checkUserHelper.js script to indicate if an XFF string is
trusted or untrusted in the summary table. Also add an image for
all the checkuser-helper-* messages in qqq.json that show
the table generated by the checkUserHelper.js script on a test wiki
to assist in translation efforts.

Bug: T329112
Change-Id: I23fd2f2e517169a413ce5d9942964f93e6ca278e