Fix summary table overflow over the page tools on Vector 2022
Specify CSS rules that prevent the table showing the summary of
the results from overlapping the Vector 22 page tools by allowing
word-breaking anywhere for text inside the "User" and "IP(s)"
columns of the table.

This occurs in testing primarily when an IPv6 address is used
to edit either anonymously or with a long username with few spaces.

To achieve this the "word-wrap: anywhere;" rule is applied to all
links inside the first two columns and a min-width value is imposed
to prevent this rule from squashing the first two columns content.
This min-width rule gives a 1/4th each to the first two columns, and
leaves a half of the width to the "User Agent(s)" column which in
most cases has much more textual content.

This minimum width rule is only applied until 1500px width for the
screen, as past this point the "User" and "IP(s)" columns will have
enough space to not look squashed.

Bug: T330031
Change-Id: I2f5a45421a6d7240a5aae7761cc1492bcd396723