Replace sonarcloud marked regex with equivalent javascript function
Replace regex to trim with functionally equivalent trim method
provided by javascript. The regex was marked as being a security
hotspot by sonarcloud for being "vulnerable to super-linear
runtime due to backtracking". Considering that the regex performed
was exactly the same as the trim method but faster and that the
trim method is supported in all browsers except IE 9 and below [1]
there is no need to continue to use the regex.

[1] https://caniuse.com/mdn-javascript_builtins_string_trim

Change-Id: I293c85fd14985ae542281113b9814f5c7463cc04