Update "Try the new CheckUser tool" link text
Update the "Try the new CheckUser tool" link text shown next to the
Help icon on Special:CheckUser with "Try out Special:Investigate"
because the tool is no longer "new". Also add a subtitle link to
Special:Investigate that includes the target on Special:CheckUser
if one is defined.

Also update the messages used in the subtitle links for
Special:Investigate and Special:CheckUser to use the name of the
special page as defined by the "#special" magic word. This is done
because the current subtitle link from Investigate could be
misinterpreted as a switching back to the CheckUser extension
instead of Special:CheckUser.

The link to Special:CheckUserLog with no target specified is changed
to "Recent checks" as this more closely matches the wording for the
link to CheckUserLog with a target specified (which is "Recent checks
for this user").

Bug: T330423
Change-Id: If4ec628a1206651bf6d6e4abcfb0d0d2242e8111