Improve wording of labels for links to Special:CheckUserLog
Improve the wording used for links to Special:CheckUserLog:
* Remove references to "recent" as the logs shown could be 10+
years old depending on how actively checked a particular target is
* Change the reference to "user" in "checkuser-recent-checks" to
"target" as this link could provide a log of checks on an XFF or
range that are unlikely to be a specific "user". The word "target"
is less specific on how many users there are.
* Remove the word "user" from "checkuser-contribs-log-initiator"
as checks shown by clicking that link could include IP address ranges
and XFF IPs that are unlikely to be specific to a "user".

The word "target" is used as the internal name of the field used to
search by the user/IP on both SpecialCheckUser and SpecialInvestigate.

Bug: T330649
Change-Id: Iaf706a3d2a0d0dd83d575ea04bd10da10b1e4ab9