Add types to the parameters of methods in HooksTest.php
Add types to the parameters of methods in HooksTest.php to make
it easier to understand what the test does. Without this someone
reading the code would need to inspect the data provider for the
type of the variables. The tests passing for this change should
ensure that this change works as intended.

Change-Id: I8d6d53e3afbdde993ecc98862a1755fbe1c717a5