Display "Show IP" buttons when using diff keywords
Currently, the "Show IP" button doesn't show on diff pages when it uses keywords (i.e. prev, cur, next) instead of a diff number. This was due to the way we checked if the page was a diff page, only considering diff pages that have a diff id. This patch just checks if diff has a value different from null.

Bug: T331007
Change-Id: Ifc6517e38238cdff8143bf8e94766e65c0992056