SECURITY: Hide hidden users in Special:CheckUserLog
Make Special:CheckUserLog hide users shown in the results list
that have been hidden if the user viewing the page does not have
the 'hideuser' right. If the user does have the 'hideuser' right
the username is struck through to indicate that the user is hidden
but the username itself is still shown.

This patch will bring an implementation that is similar to
Special:Log, where users who know the exact username that has been
hidden can find associated log entries that display "username removed"
but can never find the username if they don't already know it.

Bug: T330968
Change-Id: I5a5ca8caf2c300d431d6e561d15f58ab37ab1b2f