SECURITY: Escape HTML in the user text for the checkbox in get users
Escape any HTML in the user text provided to the mustache template.
This is needed because any double quote characters (i.e. '"') will
cause the "value" attribute for the checkbox to be ended early. By
doing this it makes blocking the account using the block form that
is shown under the results impossible, but also could allow HTML
injection into the input tag. Injection of other elements should
not be possible unless $wgInvalidUsernameCharacters was modified
to remove the ">" character.

Bug: T331192
Change-Id: If38d0ffbfed795c2f2206a09aa65fb16d6a9d030