Re-work SpecialBlock.js for UserInputWidget
The users multiselect widget has caused some problems on Special:Block
due to a bug in the widget: T334711

The user input widget is restored instead, in T334870.

For users with the right to view temporary account IP addresses, the
IPs are looked up and displayed in response to their entered target.
Update this for the user input widget:
* 'change' fires as the user types, so don't look up IP addresses,
  just clear any previous results
* instead, look up IP addresses on the menu's 'choose' event

Bug: T334870
Change-Id: I4f125565f551b5237a7aee3e3c23000607fbce3f