Only use the 'wikipage.content' hook on RecentChanges and Watchlist
RecentChanges and Watchlist require the IP reveal buttons to be added
via the wikipage.content hook, because they can continually update. For
these pages, handle the hook and scope the buttons to the passed
$content element.

For all other pages, do not listen for the hook, and scope the buttons
to the $( '#bodyContent' ) element. This allows buttons to be shown in
banners, on Special:Permalink (T331008), etc.

Factor out the common functionality to one place.

Bug: T335999
Change-Id: I2a6237a8174dbd3c115e6b70a056331811be9e03