Log when a user views IP addresses used by a temporary account
Log when a user accesses one of the temporary account handler
endpoints to view IP addresses used by a temporary account.
Enqueue a job to create the log and define formatting for the
displayed log line.

The log is debounced, so the same log line won't be created
within 24 hours.

Bug: T335366
Change-Id: I19c7cd36ed1d9793cf627438b46d1419d5dcbc17