Add tests for the CIDR calculator widget
Add tests for the CIDR calculator widget. This is not needed to
split the code, so these tests are in a seperate change to make
the first change smaller. It is also not needed for other parent
tasks to Ifd3936b0c0e58450c034a8f69a69b1d8ea82bd65.

Bug: T313291
Change-Id: Ieb8928295f2d757456fa643307ea44b111297886