Use a Job to prune data instead of post send
Use a Job to prune data from the CheckUser tables that have old
entries removed (cu_changes, cu_log_event and cu_private_event)
instead of doing this post send via a DeferredUpdate. This is
done to avoid (in the worst case) the deletion of 1,500 entries
over three tables. Doing this post send means this is run using
the server that handled the request, which has been detailed as
an issue at the Hackathon this year (2023).

Bug: T337175
Change-Id: I6447f7df532221c79c347a3cbd18fa91947f6748