Test SpecialCheckUser::tagPage and ::doMassUserBlockInternal
Test SpecialCheckUser::tagPage using unit and integration tests.
Also test ::doMassUserBlockInternal using integration testing.

While modifying tests, convert some integration tests to be unit
tests to improve the time it takes to run the tests. These tests
are for SpecialCheckUser::doesWrites and ::checkReason.

Other patches that convert integration to unit tests will be made
to improve the speed it takes to run the tests to counterbalance the
extra time spent running the integration test for
::doMassUserBlockInternal.

Change-Id: Icc1947786ca44ba3a98d21e949e7598549a23bcd