Select unaliased cuc_timestamp to allow IndexPager to use it
Select the 'cuc_timestamp' column so that the IndexPager can use
it as the timestamp field. The IndexPager cannot use the aliased
'timestamp' field because it tries to use the same field name
defined in ::getTimestampField for both the WHERE conditions and
for selecting the timestamp for the results for the Group by date
feature.

As such, selecting cuc_timestamp just for the IndexPager still
means that the CheckUser code uses the aliased fields to support
event table migration but does not affect the IndexPager.

Bug: T337639
Change-Id: Ibe1388d495f6777cf34ccb6fe64883e303a655ce