Avoid dynamic message keys via FormSpecialPage::getMessagePrefix()
This goes explicitly against the coding conventions, and the method will
be removed.

Also remove a validation callback using the
checkuser-investigate-duration-invalid message, which does not exist. Since the field is a select, it shouldn't be possible to enter an invalid value (unless you mess up with the form, I guess, but that wouldn't deserve a pretty error message).

Bug: T337945
Change-Id: I3b1d56b8dc4313289d37e5f1f3ecc757a9e40580