Reduce cognitive complexity for cidr.js updateCIDRresult()
Reduce the cognitive complexity of the updateCIDRresult() function
in cidr.js by splitting the code to get the list of IPs into a separate
function. Future patches will address the readability in other parts
of this file.

This was noted as an issue by SonarCloud and should make the code easier
to read.

This change should make no functionality changes to the CIDR calculator.

Change-Id: I416ad6001029a069dafe8af398c1f3bba9634c70