Make many tests for CheckUserUnionSelectQueryBuilder unit tests
Convert a number of tests that do not interact with the database into
unit tests. These tests cover the CheckUserUnionSelectQueryBuilder and
can be run as unit tests without any changes to the test code. This is
done to decrease the time it takes to run the tests as unit tests run
faster.

Also the single test for the factory for this class and also the tests
for a helper class are also convertible to unit tests without the need
for changes in the test code. Therefore, these are also made into unit
tests.

Change-Id: Ib8ad574e35efaf29e11677a15b1e4a06824544b8